tests: Move common thermostat checks to base class
