+++ testsuite/python/tests_common.py
@@ -645,12 +645,6 @@ class DynamicDict(dict):
         return eval(value, self) if isinstance(value, str) else value
 
 
-def single_component_maxwell(x1, x2, kT):
-    """Integrate the probability density from x1 to x2 using the trapezoidal rule"""
-    x = np.linspace(x1, x2, 1000)
-    return np.trapz(np.exp(-x**2 / (2. * kT)), x) / np.sqrt(2. * np.pi * kT)
-
-
 def count_fluid_nodes(lbf):
     """Counts the non-boundary nodes in the passed lb fluid instance."""
